pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-board-games"
version = "0.1.0"
description = "Reinforcement learning agents for board games, starting with Ricochet Robots"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "RL Board Games Team"}
]
dependencies = [
    "gymnasium>=0.28",
    "stable-baselines3>=2.1.0",
    "torch>=2.0",
    "numpy>=1.23",
    "numba>=0.57",
    "networkx>=3.1",
    "wandb>=0.16",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-benchmark>=4.0",
    "hypothesis>=6.88",
    "black>=23.7",
    "isort>=5.12",
    "flake8>=6.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["rl_board_games*"]

[tool.black]
line-length = 120
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 120 